requirements.txt
pynetbox>=6.6.2,<7.0
click>=8.1.3,<9.0
netaddr
Jinja2
python-dotenv
dnspython
